Add files via upload
SmartFingerCounter is a sophisticated hand-tracking system designed to detect and count the number of fingers raised on one or two hands using a webcam. Utilizing OpenCV for image processing and Mediapipe for accurate hand landmark detection, this project provides real-time feedback on finger counts displayed on the screen. It’s ideal for applications in gesture recognition, interactive user interfaces, and various hand-tracking tasks. The system can be easily extended for additional functionalities, such as gesture-based commands or integration with other technologies.